+++ src/Settings/RemoveWallet/RemoveWalletScreen.tsx
@@ -6,15 +6,15 @@ import {SafeAreaView} from 'react-native-safe-area-context'
 import {useMutation, UseMutationOptions} from 'react-query'
 import {useDispatch, useSelector} from 'react-redux'
 
-import {updateWallets} from '../../legacy/actions'
-import styles from '../../legacy/components/Settings/styles/RemoveWalletScreen.style'
-import {Button, Checkbox, Spacer, StatusBar, Text, TextInput} from '../../legacy/components/UiKit'
-import {Checkmark} from '../../legacy/components/UiKit/TextInput'
-import walletManager from '../../legacy/crypto/walletManager'
-import {WALLET_ROOT_ROUTES} from '../../legacy/RoutesList'
-import {isHWSelector} from '../../legacy/selectors'
-import {useWalletName} from '../hooks'
-import {useSelectedWallet} from '../SelectedWallet'
+import {updateWallets} from '../../../legacy/actions'
+import styles from '../../../legacy/components/Settings/styles/RemoveWalletScreen.style'
+import {Button, Checkbox, Spacer, StatusBar, Text, TextInput} from '../../../legacy/components/UiKit'
+import {Checkmark} from '../../../legacy/components/UiKit/TextInput'
+import walletManager from '../../../legacy/crypto/walletManager'
+import {WALLET_ROOT_ROUTES} from '../../../legacy/RoutesList'
+import {isHWSelector} from '../../../legacy/selectors'
+import {useWalletName} from '../../hooks'
+import {useSelectedWallet} from '../../SelectedWallet'
 
 export const RemoveWalletScreen = () => {
   const strings = useStrings()
